+++ src/core/callbacks.rs
@@ -65,4 +65,9 @@ pub enum CallbackEvent {
         widget_id: i32,
         button: Button,
     },
+
+    WidgetClicked {
+        widget_id: i32,
+        button: Button,
+    }
 }